refactor(api): split handler.go into per-resource handler files
Moved all structs, handlers, and helpers from the monolithic handler.go
(1832 lines) into focused files: routes.go, handler_agents.go,
handler_offices.go, handler_schedules.go, handler_memories.go,
handler_skills.go, handler_secrets.go, handler_connectors.go,
handler_templates.go. No logic changes.
